refactor: Reorganize filenames, Battleship => Turn, Turn => Game
